+++ cleo_plugins/MemoryOperations/MemoryOperations.cpp
@@ -3,8 +3,7 @@
 #include "plugin.h"
 #include "CTheScripts.h"
 #include <filesystem>
-#include <map>
-#include <set>
+#include <unordered_map>
 
 using namespace CLEO;
 using namespace plugin;
@@ -12,12 +11,22 @@ using namespace plugin;
 class MemoryOperations 
 {
 public:
-    static std::set<void*> m_allocations;
-    static std::map<HMODULE, size_t> m_libraries;
+    std::unordered_map<void*, size_t> m_allocations;
+    std::unordered_map<HMODULE, size_t> m_libraries;
+
+    // keep track of per-script memory allocations
+    struct AllocationInfo { int count; int size; };
+    std::unordered_map<CLEO::CRunningScript*, AllocationInfo> m_scriptAllocationsInfo;
+    int m_configLimitAllocationCount;
+    int m_configLimitAllocationSize;
 
     MemoryOperations()
     {
         if (!PluginCheckCleoVersion()) return;
+
+        auto config = GetConfigFilename();
+        m_configLimitAllocationCount = GetPrivateProfileInt("Limits", "MemoryAllocations", 2000, config.c_str());
+        m_configLimitAllocationSize = GetPrivateProfileInt("Limits", "MemoryTotalSize", 16, config.c_str()) * 1024 * 1024; // megabytes
 
         //register opcodes
         CLEO_RegisterOpcode(0x0459, opcode_0459); // terminate_all_scripts_with_this_name
@@ -65,31 +74,71 @@ public:
         
 
         // register event callbacks
-        CLEO_RegisterCallback(eCallbackId::ScriptsFinalize, OnFinalizeScriptObjects);
+        CLEO_RegisterCallback(eCallbackId::GameEnd, OnGameEnd);
     }
 
     ~MemoryOperations()
     {
-        CLEO_UnregisterCallback(eCallbackId::ScriptsFinalize, OnFinalizeScriptObjects);
+        CLEO_UnregisterCallback(eCallbackId::GameEnd, OnGameEnd);
     }
 
-    static void __stdcall OnFinalizeScriptObjects()
+    static void __stdcall OnGameEnd()
     {
-        TRACE("Cleaning up %d allocated memory blocks...", m_allocations.size());
-        for (auto p : m_allocations) free(p);
-        m_allocations.clear();
-
-        size_t libCount = std::count_if(m_libraries.begin(), m_libraries.end(), [](auto& entry) { return entry.second; });
-        TRACE("Cleaning up %d loaded libraries...", libCount);
-        for (auto& entry : m_libraries)
+        // release memory allocations
+        TRACE("");
+        TRACE("Cleaning up %d allocated memory block(s):", Instance.m_allocations.size());
+        for (auto entry : Instance.m_scriptAllocationsInfo) // list remaining allocations per script
         {
+            if (entry.second.count == 0) continue;
+
+            std::string str(128, '\0');
+            CLEO_GetScriptInfoStr(entry.first, false, str.data(), str.length());
+            TRACE(" %d block%s (%0.2f kB) in script %s",
+                entry.second.count,
+                entry.second.count > 1 ? "s" : "",
+                float(entry.second.size) / 1024,
+                str.c_str());
+        }
+        for (auto p : Instance.m_allocations) free(p.first);
+        Instance.m_allocations.clear();
+        Instance.m_scriptAllocationsInfo.clear();
+
+        // release loaded dlls
+        size_t libCount = std::count_if(Instance.m_libraries.begin(), Instance.m_libraries.end(), [](auto& entry) { return entry.second; });
+        TRACE("");
+        TRACE("Cleaning up %d loaded libraries:", libCount);
+        for (auto& entry : Instance.m_libraries)
+        {
+            if (entry.second == 0) continue;
+
+            std::string str(MAX_PATH, '\0');
+            GetModuleFileNameA(entry.first, str.data(), str.length());
+            FilepathRemoveParent(str, CLEO_GetGameDirectory());
+            TRACE(" %s", str.c_str());
+
             while (entry.second > 0)
             {
                 FreeLibrary(entry.first);
                 entry.second -= 1;
             }
         }
-        m_libraries.clear();
+        Instance.m_libraries.clear();
+    }
+
+    void RegisterMemoryAllocation(CLEO::CRunningScript* thread, void* address, size_t size)
+    {
+        m_allocations[address] = size;
+        auto& info = m_scriptAllocationsInfo[thread];
+        info.count++;
+        info.size += size;
+    }
+
+    void UnregisterMemoryAllocation(CLEO::CRunningScript* thread, void* address)
+    {
+        auto& info = m_scriptAllocationsInfo[thread];
+        info.count--;
+        info.size -= m_allocations[address];
+        m_allocations.erase(address);
     }
 
     // opcodes 0AA5 - 0AA8
@@ -441,7 +490,7 @@ public:
 
         if (ptr != nullptr)
         {
-            m_libraries[ptr] += 1;
+            Instance.m_libraries[ptr] += 1;
         }
 
         OPCODE_WRITE_PARAM_PTR(ptr);
@@ -455,20 +504,20 @@ public:
         auto ptr = (HMODULE)OPCODE_READ_PARAM_PTR();
 
         // validate
-        if (m_libraries.find(ptr) == m_libraries.end())
+        if (Instance.m_libraries.find(ptr) == Instance.m_libraries.end())
         {
             LOG_WARNING(thread, "Invalid '0x%X' library pointer param in script %s", ptr, ScriptInfoStr(thread).c_str());
             return OR_CONTINUE;
         }
 
-        if (m_libraries[ptr] == 0)
+        if (Instance.m_libraries[ptr] == 0)
         {
             LOG_WARNING(thread, "Trying to free already unloaded library in script %s", ScriptInfoStr(thread).c_str());
             return OR_CONTINUE;
         }
 
         FreeLibrary(ptr);
-        m_libraries[ptr] -= 1;
+        Instance.m_libraries[ptr] -= 1;
         return OR_CONTINUE;
     }
 
@@ -645,7 +694,17 @@ public:
             DWORD oldProtect;
             VirtualProtect(mem, size, PAGE_EXECUTE_READWRITE, &oldProtect);
 
-            m_allocations.insert(mem);
+            Instance.RegisterMemoryAllocation(thread, mem, size);
+
+            const auto& info = Instance.m_scriptAllocationsInfo[thread];
+            if (Instance.m_configLimitAllocationSize > 0 && info.size > Instance.m_configLimitAllocationSize)
+            {
+                LOG_WARNING(thread, "%d MB of memory currently allocated by script %s", info.size / (1024 * 1024), ScriptInfoStr(thread).c_str());
+            }
+            else if (Instance.m_configLimitAllocationCount > 0 && info.count > Instance.m_configLimitAllocationCount)
+            {
+                LOG_WARNING(thread, "More than %d memory blocks currently allocated by script %s", Instance.m_configLimitAllocationCount, ScriptInfoStr(thread).c_str());
+            }
         }
         else
             LOG_WARNING(thread, "Failed to allocate %d bytes of memory in script %s", size, ScriptInfoStr(thread).c_str());
@@ -662,14 +721,16 @@ public:
         auto address = OPCODE_READ_PARAM_PTR();
 
         // validate params
-        if (m_allocations.find(address) == m_allocations.end())
+        if (Instance.m_allocations.find(address) == Instance.m_allocations.end())
         {
             LOG_WARNING(thread, "Invalid '0x%X' pointer param to unknown or already freed memory in script %s", address, ScriptInfoStr(thread).c_str());
             return OR_CONTINUE;
         }
 
         free(address);
-        m_allocations.erase(address);
+        
+        Instance.UnregisterMemoryAllocation(thread, address);
+
         return OR_CONTINUE; // done
     }
 
@@ -858,13 +919,14 @@ public:
         auto address = OPCODE_READ_PARAM_PTR();
 
         // validate params
-        if (m_allocations.find(address) == m_allocations.end())
+        if (Instance.m_allocations.find(address) == Instance.m_allocations.end())
         {
             LOG_WARNING(thread, "Invalid '0x%X' pointer param to unknown or already freed memory in script %s", address, ScriptInfoStr(thread).c_str());
             return OR_CONTINUE;
         }
 
-        m_allocations.erase(address);
+        Instance.UnregisterMemoryAllocation(thread, address);
+
         return OR_CONTINUE; // done
     }
 
@@ -937,7 +999,5 @@ public:
 
         return (address == thread) ? OR_INTERRUPT : OR_CONTINUE;
     }
-} Memory;
+} Instance;
 
-std::set<void*> MemoryOperations::m_allocations;
-std::map<HMODULE, size_t> MemoryOperations::m_libraries;